fix bug salary

+++ Jobs-board-backend/DB/db_setup.sql
@@ -51,8 +51,8 @@ CREATE TABLE IF NOT EXISTS jobs (
 ALTER TABLE jobs
   ADD COLUMN IF NOT EXISTS location_type VARCHAR(50),
   ADD COLUMN IF NOT EXISTS api_source TEXT DEFAULT 'DevitJobs',
-  ADD COLUMN IF NOT EXISTS salary_min NUMERIC,
-  ADD COLUMN IF NOT EXISTS salary_max NUMERIC;
+  ADD COLUMN IF NOT EXISTS salary_min INT,
+  ADD COLUMN IF NOT EXISTS salary_max INT;
 
 -- application_status enum + applications table
 DO $$ BEGIN
